requirements.txt
flask
torch
transformers
sentencepiece
gdown
psycopg2-binary>=2.9.9
